Edits to allow specal Massachussets state rules when calcualting unit cost This should probably be revisited as I suspect other states may have differing laws about how unit prices should be reported and right now the way unit price is caluclated is a bit of a mess.
Mostly it's taken from PriceLib which is problematic mecause it's not easy to override
unless I'm missing something. But it is also cacualted in the tag data sources.

Except the tag data sources wren't in use and everything was just being calculated out in
the products model for the getTagData() function.

I've made it so that the system will always use a tag data source and added a default that
does what the function in the products model used to do. I also made the EditBatchPage genereate
tags based on the tag data source instead of doing it's own thing.

My soultion for the PriceLib issues is super awful rightnow because I'm on a deadline so I've
pepperd the code with ifs based on my FANNIE_COOP_ID that I'll need to find a way to remove when I
get a minute.